Mostly hopefully finished core stuff?
A lot of what I wrote I think in retrospect was probably more comlicated
than it might have had to have been. Still, I got it working through at
least the little bit of testing that I did in the tester.c file. Feel
free to write more tests and try to break it. The tester.c file is
waiting to be added to.