Did read only and writeByte support
Completed the read only and writeByte() part of the assignment. Added
read/write byte to the indexblock structure. Also added an index block
field to the filedata structure since it'll be useful for defrag. Added
tests to tester.c to confirm that the additions work at least in a
minimal sense.